refact: ソーシャルリンクに関する関数をutilsに移動
